add declaration of solid and Euler operations

+++ src/Vertex.h
@@ -5,8 +5,8 @@
 class Vertex
 {
 public:
-    float x, y, z; // coordinates of the vertex
+    double x, y, z; // coordinates of the vertex
     HalfEdge *edge; // starting point of a half edge
 
-    Vertex(float x, float y, float z) : x(x), y(y), z(z), edge(nullptr) {}
+    Vertex(double x, double y, double z) : x(x), y(y), z(z), edge(nullptr) {}
 };
